requirements.txt
discord.py
Flask
aiosqlite
